containers/piercedVector: fix 'pierce_pos' when piercing the last position
When piercing the last position, the position that need to be
marked as empty is the position after the last one.

There is no need to explicitly mark as empty the current position,
because that position can now be after the sentinel element. If the
current position is now the new sentinel, it will be correctly marked
as such, because it will be the position after the last one.