Modify file python_wrapper/phd_thesis/project/Laplacian2D.py
Solved a frustrating bug: just to level 7 of refinement everything worked
fine, but going up of one level gave us strange behaviors in the solution.

The problem was in the input data of function "get_point_owner_idx", that
in the nex "bitpit" implementation wants an array C++ of 3 elements.

Passing into a list of 2 elements, gave these unexpected problems.