"init_mat" now has no more "o_n_oct" as parameter (was useless)
Plus, deleted check for "mapping" (we want to be always in mapping config)
and applied a bit of code hoisting.

Deleted  some useless variable and rewritten some parts to better
understand the function.

Commented "w_first.h_c_w_first(.....)" because it was already considered
into the computation of "metric_coefficients". That's why "w_first" is
set equal to 1.0.