Deleted check for "mapping" (we want to be always in mapping config) and applied a bit of code hoisting and lambda functions in functions "update_bg_grids" and "update_fg_grids".
Deleted  some useless variable or commands and rewritten some parts to
better understand the function.

Plus, deleted "utilities.h_c_w_first(.....)" and "n_ws_first" because it
was already considered into the computation of "metric_coefficients", and
it was set equal to 1.0. That's why it was useless inside the operations.